pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "greenroom"
version = "0.1.0"
description = "Greenroom - MCP server for entertainment content recommendations"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Chris Brickey", email = "chrisbrickey@example.com"}
]
dependencies = [
    "fastmcp>=2.13.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
greenroom = "greenroom.server:main"

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.35.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/greenroom"]
